+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/text/ExoplayerCuesDecoder.java
@@ -86,6 +86,11 @@ public final class ExoplayerCuesDecoder implements SubtitleDecoder {
     return "ExoplayerCuesDecoder";
   }
 
+  @Override
+  public void setOutputStartTimeUs(long outputStartTimeUs) {
+    // Do nothing.
+  }
+
   @Nullable
   @Override
   public SubtitleInputBuffer dequeueInputBuffer() throws SubtitleDecoderException {