+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/video/DecoderVideoRenderer.java
@@ -366,7 +366,9 @@ public abstract class DecoderVideoRenderer extends BaseRenderer {
         outputBuffer.release();
         outputBuffer = null;
       }
-      checkNotNull(decoder).flush();
+      Decoder<?, ?, ?> decoder = checkNotNull(this.decoder);
+      decoder.flush();
+      decoder.setOutputStartTimeUs(getLastResetPositionUs());
       decoderReceivedBuffers = false;
     }
   }
@@ -718,6 +720,7 @@ public abstract class DecoderVideoRenderer extends BaseRenderer {
     try {
       long decoderInitializingTimestamp = SystemClock.elapsedRealtime();
       decoder = createDecoder(checkNotNull(inputFormat), cryptoConfig);
+      decoder.setOutputStartTimeUs(getLastResetPositionUs());
       setDecoderOutputMode(outputMode);
       long decoderInitializedTimestamp = SystemClock.elapsedRealtime();
       eventDispatcher.decoderInitialized(